output: split otlp module into otlp_http and otlp_grpc
Replaces the single `output otlp { protocol grpc | http_* }` DSL type
with two transport-specific modules (`output otlp_http` and `output
otlp_grpc`), mirroring the input side which has had split modules
since 0.7.0. Eliminates the `protocol`-keyed conditional surface that
touched headers / metadata semantics, `verify`, endpoint conventions,
and other transport-specific bits.

Wire-level behaviour is byte-equivalent. The protobuf encoding, retry
semantics, batch_level merging, and TLS surface all carry over. Shared
bits (BatchLevel, merge_by_*, retry block schema, OtlpPayload,
decode_drained_to_request) live under output/otlp/mod.rs as internal
helpers; the public modules are output/otlp/{http,grpc}.rs.

Breaking: `type otlp` is no longer registered. Configs must migrate to
`type otlp_http` or `type otlp_grpc`. See CHANGELOG 0.7.6 for the
migration table.

This is a pure refactor; per-peer rotation and mTLS for these modules
land in a follow-up commit on the same branch.
