+++ docs/src/SUMMARY.md
@@ -28,7 +28,8 @@
   - [syslog_udp](./outputs/syslog-udp.md)
   - [unix_socket](./outputs/unix-socket.md)
   - [stdout](./outputs/stdout.md)
-  - [otlp](./outputs/otlp.md)
+  - [otlp_http](./outputs/otlp_http.md)
+  - [otlp_grpc](./outputs/otlp_grpc.md)
 - [Processing](./processing/README.md)
   - [User-defined Processes](./processing/user-defined.md)
   - [Process Design Guide](./processing/design-guide.md)
